TCP_tester working 4 messages
