+++ tcp_tester.py
@@ -10,7 +10,7 @@ test_handshake_0 = b"\x00\x03U\xc7\xaa\xa3\x85\xe8%\x95M\x96\xcbQ\x80C\x04\x0f\x
 test_handshake_1 = b"\x00\x02L\xdf\xd9\x81\x98\xcfr\xd8\xa7d\xd2\x167\x98\xff\x9b\t\x16\x1cR\x82^\x96\t8\xfb[\x9fv\x15d\n\xc0\xf7Wi\xf2\x1f\x9f\xd6ht\xba.\xf0>\\\x1c"
 test_handshake_2 = b"\x00\xae);;\xcd\x02\xea\x12A\xfc@\xb6L\xd6\xd2.\x8by\xfc\xddIR\xd9\x9e\x86\x96j\xbf\x8cA\xec\x8aD\xb0\xf1\xcb\xd6\xedQzq\xc3,\xb3W_\xf25\x0b\x066j\xd7\x06\xd3\xa0\xf0i=\xcd\xd8J\xb0\xffv"
 test_handshake_3 = b'\x00-\x00\x10\x00\x02"\x00\x00\x03\x02\xaa\xa2\x01 \x06"nF\x11\x1a\x0bY\xca\xaf\x12`C\xeb[\xbf(\xc3O:^3*\x1f\xc7\xb2\xb7<\xf1\x88\x91\x0f'
-handshakes = [test_handshake_0] #, test_handshake_1, test_handshake_2, test_handshake_3]
+handshakes = [test_handshake_0, test_handshake_1, test_handshake_2, test_handshake_3]
 
 
 async def send_raw(proxy_port, packets):
@@ -19,7 +19,7 @@ async def send_raw(proxy_port, packets):
         print("sending data:", data)
         await proxy_port.send_all(data)
         # space out packets
-        await trio.sleep(0.1)
+        await trio.sleep(50)
 
 
 async def tester_client(output_stream):
@@ -38,7 +38,7 @@ async def tester_client(output_stream):
 
 async def tester_server(input_stream):
     print("fldigi input server started")
-    delay = 2
+    delay = 5
     try:
         print("Sleeping for", delay, "seconds")
         await trio.sleep(delay)
